fix bench example ++ use config defaults in the benches
